moved todoBlock instances creating to TodoBlockFactory
