+++ src/test/java/com/knighttodo/knighttodo/factories/TodoFactory.java
@@ -7,9 +7,6 @@ import com.knighttodo.knighttodo.gateway.privatedb.representation.enums.Hardness
 import com.knighttodo.knighttodo.gateway.privatedb.representation.enums.Scariness;
 import com.knighttodo.knighttodo.rest.dto.todo.request.CreateTodoRequestDto;
 import com.knighttodo.knighttodo.rest.dto.todo.request.UpdateTodoRequestDto;
-import com.knighttodo.knighttodo.rest.dto.todoblock.request.CreateTodoBlockRequestDto;
-import com.knighttodo.knighttodo.rest.dto.todoblock.request.UpdateTodoBlockRequestDto;
-import java.util.ArrayList;
 
 public class TodoFactory {
 
@@ -22,9 +19,6 @@ public class TodoFactory {
     public static final Integer HARD_SCARY_EXPERIENCE = 37;
     public static final Scariness UPDATED_SCARINESS_TODO = Scariness.SCARY;
     public static final Hardness UPDATED_HARDNESS_TODO = Hardness.HARD;
-
-    public static final String BLOCK_NAME = "Sunday Todos";
-    public static final String UPDATED_BLOCK_NAME = "Friday Todos";
 
     private TodoFactory() {
     }
@@ -115,65 +109,6 @@ public class TodoFactory {
     public static UpdateTodoRequestDto updateTodoRequestDtoWithoutHardness(TodoBlock savedTodoBlock) {
         UpdateTodoRequestDto request = updateTodoRequestDto(savedTodoBlock);
         request.setHardness(null);
-        return request;
-    }
-
-    public static CreateTodoBlockRequestDto createTodoBlockRequestDto() {
-        return CreateTodoBlockRequestDto
-            .builder()
-            .blockName(BLOCK_NAME)
-            .build();
-    }
-
-    public static CreateTodoBlockRequestDto createTodoBlockRequestDtoWithoutName() {
-        CreateTodoBlockRequestDto request = createTodoBlockRequestDto();
-        request.setBlockName(null);
-        return request;
-    }
-
-    public static CreateTodoBlockRequestDto createTodoBlockRequestDtoWithNameConsistingOfSpaces() {
-        CreateTodoBlockRequestDto request = createTodoBlockRequestDto();
-        request.setBlockName("    ");
-        return request;
-    }
-
-    public static TodoBlock todoBlockInstance() {
-        return TodoBlock
-            .builder()
-            .blockName(BLOCK_NAME)
-            .todos(new ArrayList<>())
-            .build();
-    }
-
-    public static UpdateTodoBlockRequestDto updateTodoBlockRequestDto(TodoBlock todoBlock) {
-        return UpdateTodoBlockRequestDto
-            .builder()
-            .id(todoBlock.getId())
-            .blockName(UPDATED_BLOCK_NAME)
-            .build();
-    }
-
-    public static UpdateTodoBlockRequestDto updateTodoBlockRequestDtoWithoutId(TodoBlock todoBlock) {
-        UpdateTodoBlockRequestDto request = updateTodoBlockRequestDto(todoBlock);
-        request.setId(null);
-        return request;
-    }
-
-    public static UpdateTodoBlockRequestDto updateTodoBlockRequestDtoWithIdConsistingOfSpaces(TodoBlock todoBlock) {
-        UpdateTodoBlockRequestDto request = updateTodoBlockRequestDto(todoBlock);
-        request.setId("    ");
-        return request;
-    }
-
-    public static UpdateTodoBlockRequestDto updateTodoBlockRequestDtoWithoutName(TodoBlock todoBlock) {
-        UpdateTodoBlockRequestDto request = updateTodoBlockRequestDto(todoBlock);
-        request.setBlockName(null);
-        return request;
-    }
-
-    public static UpdateTodoBlockRequestDto updateTodoBlockRequestDtoWithNameConsistingOfSpaces(TodoBlock todoBlock) {
-        UpdateTodoBlockRequestDto request = updateTodoBlockRequestDto(todoBlock);
-        request.setBlockName("    ");
         return request;
     }
