+++ src/test/java/com/knighttodo/knighttodo/integration/TodoBlockResourceIntegrationTest.java
@@ -15,7 +15,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.knighttodo.knighttodo.factories.TodoFactory;
+import com.knighttodo.knighttodo.factories.TodoBlockFactory;
 import com.knighttodo.knighttodo.gateway.privatedb.repository.TodoBlockRepository;
 import com.knighttodo.knighttodo.gateway.privatedb.representation.TodoBlock;
 import com.knighttodo.knighttodo.rest.dto.todoblock.request.CreateTodoBlockRequestDto;
@@ -52,7 +52,7 @@ public class TodoBlockResourceIntegrationTest {
 
     @Test
     public void addTodoBlock_shouldAddTodoBlockAndReturnIt_whenRequestIsCorrect() throws Exception {
-        CreateTodoBlockRequestDto requestDto = TodoFactory.createTodoBlockRequestDto();
+        CreateTodoBlockRequestDto requestDto = TodoBlockFactory.createTodoBlockRequestDto();
 
         mockMvc.perform(
             post(API_BASE_BLOCKS)
@@ -66,7 +66,7 @@ public class TodoBlockResourceIntegrationTest {
 
     @Test
     public void addTodoBlock_shouldRespondWithBadRequestStatus_whenNameIsNull() throws Exception {
-        CreateTodoBlockRequestDto requestDto = TodoFactory.createTodoBlockRequestDtoWithoutName();
+        CreateTodoBlockRequestDto requestDto = TodoBlockFactory.createTodoBlockRequestDtoWithoutName();
 
         mockMvc.perform(
             post(API_BASE_BLOCKS)
@@ -79,7 +79,7 @@ public class TodoBlockResourceIntegrationTest {
 
     @Test
     public void addTodoBlock_shouldRespondWithBadRequestStatus_whenNameConsistsOfSpaces() throws Exception {
-        CreateTodoBlockRequestDto requestDto = TodoFactory.createTodoBlockRequestDtoWithNameConsistingOfSpaces();
+        CreateTodoBlockRequestDto requestDto = TodoBlockFactory.createTodoBlockRequestDtoWithNameConsistingOfSpaces();
 
         mockMvc.perform(
             post(API_BASE_BLOCKS)
@@ -92,8 +92,8 @@ public class TodoBlockResourceIntegrationTest {
 
     @Test
     public void getAllTodoBlocks_shouldReturnAllTodoBlocks() throws Exception {
-        todoBlockRepository.save(TodoFactory.todoBlockInstance());
-        todoBlockRepository.save(TodoFactory.todoBlockInstance());
+        todoBlockRepository.save(TodoBlockFactory.todoBlockInstance());
+        todoBlockRepository.save(TodoBlockFactory.todoBlockInstance());
 
         mockMvc.perform(
             get(API_BASE_BLOCKS))
@@ -103,7 +103,7 @@ public class TodoBlockResourceIntegrationTest {
 
     @Test
     public void getTodoBlockById_shouldReturnExistingTodoBlock_whenIdIsCorrect() throws Exception {
-        TodoBlock todoBlock = todoBlockRepository.save(TodoFactory.todoBlockInstance());
+        TodoBlock todoBlock = todoBlockRepository.save(TodoBlockFactory.todoBlockInstance());
 
         mockMvc.perform(
             get(buildGetBlockByIdUrl(todoBlock.getId())))
@@ -113,8 +113,8 @@ public class TodoBlockResourceIntegrationTest {
 
     @Test
     public void updateTodoBlock_shouldUpdateTodoBlockAndReturnIt_whenRequestIsCorrect() throws Exception {
-        TodoBlock todoBlock = todoBlockRepository.save(TodoFactory.todoBlockInstance());
-        UpdateTodoBlockRequestDto requestDto = TodoFactory.updateTodoBlockRequestDto(todoBlock);
+        TodoBlock todoBlock = todoBlockRepository.save(TodoBlockFactory.todoBlockInstance());
+        UpdateTodoBlockRequestDto requestDto = TodoBlockFactory.updateTodoBlockRequestDto(todoBlock);
 
         mockMvc.perform(
             put(API_BASE_BLOCKS)
@@ -129,8 +129,8 @@ public class TodoBlockResourceIntegrationTest {
 
     @Test
     public void updateTodoBlock_shouldRespondWithBadRequestStatus_whenIdIsNull() throws Exception {
-        TodoBlock todoBlock = todoBlockRepository.save(TodoFactory.todoBlockInstance());
-        UpdateTodoBlockRequestDto requestDto = TodoFactory.updateTodoBlockRequestDtoWithoutId(todoBlock);
+        TodoBlock todoBlock = todoBlockRepository.save(TodoBlockFactory.todoBlockInstance());
+        UpdateTodoBlockRequestDto requestDto = TodoBlockFactory.updateTodoBlockRequestDtoWithoutId(todoBlock);
 
         mockMvc.perform(put(API_BASE_BLOCKS)
                             .content(objectMapper.writeValueAsString(requestDto))
@@ -140,8 +140,8 @@ public class TodoBlockResourceIntegrationTest {
 
     @Test
     public void updateTodoBlock_shouldRespondWithBadRequestStatus_whenIdConsistsOfSpaces() throws Exception {
-        TodoBlock todoBlock = todoBlockRepository.save(TodoFactory.todoBlockInstance());
-        UpdateTodoBlockRequestDto requestDto = TodoFactory.updateTodoBlockRequestDtoWithIdConsistingOfSpaces(todoBlock);
+        TodoBlock todoBlock = todoBlockRepository.save(TodoBlockFactory.todoBlockInstance());
+        UpdateTodoBlockRequestDto requestDto = TodoBlockFactory.updateTodoBlockRequestDtoWithIdConsistingOfSpaces(todoBlock);
 
         mockMvc.perform(
             put(API_BASE_BLOCKS)
@@ -152,8 +152,8 @@ public class TodoBlockResourceIntegrationTest {
 
     @Test
     public void updateTodoBlock_shouldRespondWithBadRequestStatus_whenBlockNameIsNull() throws Exception {
-        TodoBlock todoBlock = todoBlockRepository.save(TodoFactory.todoBlockInstance());
-        UpdateTodoBlockRequestDto requestDto = TodoFactory.updateTodoBlockRequestDtoWithoutName(todoBlock);
+        TodoBlock todoBlock = todoBlockRepository.save(TodoBlockFactory.todoBlockInstance());
+        UpdateTodoBlockRequestDto requestDto = TodoBlockFactory.updateTodoBlockRequestDtoWithoutName(todoBlock);
 
         mockMvc.perform(
             put(API_BASE_BLOCKS)
@@ -166,8 +166,8 @@ public class TodoBlockResourceIntegrationTest {
 
     @Test
     public void updateTodoBlock_shouldRespondWithBadRequestStatus_whenBlockNameConsistsOfSpaces() throws Exception {
-        TodoBlock todoBlock = todoBlockRepository.save(TodoFactory.todoBlockInstance());
-        UpdateTodoBlockRequestDto requestDto = TodoFactory
+        TodoBlock todoBlock = todoBlockRepository.save(TodoBlockFactory.todoBlockInstance());
+        UpdateTodoBlockRequestDto requestDto = TodoBlockFactory
             .updateTodoBlockRequestDtoWithNameConsistingOfSpaces(todoBlock);
 
         mockMvc.perform(
@@ -182,7 +182,7 @@ public class TodoBlockResourceIntegrationTest {
     @Test
     @Transactional
     public void deleteTodoBlock_shouldDeleteTodoBlock_whenIdIsCorrect() throws Exception {
-        TodoBlock todoBlock = todoBlockRepository.save(TodoFactory.todoBlockInstance());
+        TodoBlock todoBlock = todoBlockRepository.save(TodoBlockFactory.todoBlockInstance());
 
         mockMvc.perform(
             delete(buildDeleteBlockByIdUrl(todoBlock.getId())))
